+++ rtos/TARGET_CORTEX/rtx5/rt_OsEventObserver.c
