+++ rtos/TARGET_CORTEX/rtx5/rtx_lib.c
