+++ rtos/TARGET_CORTEX/rtx5/rtx_memory.c
